Formatting + code style adjustments
