+++ mobile/app/(tabs)/profile/passport.tsx
@@ -11,157 +11,156 @@ interface Event {
   date: string;
 }
 
-export default function FilterPassport() {
+export default function Passport() {
   const user = useSelector((state: any) => { return { userId: state.auth.userId, authToken: state.auth.authToken } });
   const [userEventDetails, setUserEventsDetails] = useState<string[]>([]);
   const [events, setEvents] = useState<Event[]>([]);
   const [open, setOpen] = useState(false);
   const [isEventsFetched, setIsEventsFetched] = useState(false);
   const [items, setItems] = useState([
-    {label: 'Most Recent', value: 'mostRecent'},
-    {label: 'Least Recent', value: 'leastRecent'},
-]);
+    { label: 'Most Recent', value: 'mostRecent' },
+    { label: 'Least Recent', value: 'leastRecent' },
+  ]);
   const [sortOrder, setSortOrder] = useState<'mostRecent' | 'leastRecent'>('mostRecent');
 
   useEffect(() => {
-      fetchUserEventDetails();
+    fetchUserEventDetails();
   }, []);
 
-   useEffect(() => {
+  useEffect(() => {
     if (isEventsFetched) {
       sortEvents(sortOrder);
     }
-  }, [isEventsFetched, sortOrder]); 
+  }, [isEventsFetched, sortOrder]);
 
   useEffect(() => {
-      sortEvents(sortOrder);
+    sortEvents(sortOrder);
   }, [sortOrder]);
 
   const fetchUserEventDetails = async () => {
     console.log("in fetch user event details");
-      try {
-        console.log(`http://${process.env.EXPO_PUBLIC_SERVER_IP}:${process.env.EXPO_PUBLIC_SERVER_PORT}/users/${user.userId}`)
-          const response = await axios.get(`http://${process.env.EXPO_PUBLIC_SERVER_IP}:${process.env.EXPO_PUBLIC_SERVER_PORT}/users/${user.userId}`,
-              {
-                  headers: {
-                      'Authorization': `Bearer ${user.authToken}`,
-                      'Content-Type': 'application/json'
-                  },
-              }
-          );
-          setUserEventsDetails(response.data.data.registeredEvents || []);
-          const eventIDs = response.data.data.registeredEvents;
-          let eventData = [];
-          for (let i = 0; i < eventIDs.length; i++)
-          {
-            const event = await fetchUserEvent(eventIDs[i]);
-            if (event !== null)
-            {
-              eventData.push(event);
-            }
-          }
-          setEvents(eventData);
-          setIsEventsFetched(true);
-      } catch (err) {
-          console.error('Error fetching events:', err);
+    try {
+      console.log(`http://${process.env.EXPO_PUBLIC_SERVER_IP}:${process.env.EXPO_PUBLIC_SERVER_PORT}/users/${user.userId}`)
+      const response = await axios.get(`http://${process.env.EXPO_PUBLIC_SERVER_IP}:${process.env.EXPO_PUBLIC_SERVER_PORT}/users/${user.userId}`,
+        {
+          headers: {
+            'Authorization': `Bearer ${user.authToken}`,
+            'Content-Type': 'application/json'
+          },
+        }
+      );
+      setUserEventsDetails(response.data.data.registeredEvents || []);
+      const eventIDs = response.data.data.registeredEvents;
+      let eventData = [];
+      for (let i = 0; i < eventIDs.length; i++) {
+        const event = await fetchUserEvent(eventIDs[i]);
+        if (event !== null) {
+          eventData.push(event);
+        }
       }
+      setEvents(eventData);
+      setIsEventsFetched(true);
+    } catch (err) {
+      console.error('Error fetching events:', err);
+    }
   };
 
   const fetchUserEvent = async (eventId: string) => {
     try {
-        const response = await axios.get(`http://${process.env.EXPO_PUBLIC_SERVER_IP}:${process.env.EXPO_PUBLIC_SERVER_PORT}/events/${eventId}`,
-            {
-                headers: {
-                    'Authorization': `Bearer ${user.authToken}`,
-                    'Content-Type': 'application/json'
-                },
-            }
-        );
-        return response.data.data;
+      const response = await axios.get(`http://${process.env.EXPO_PUBLIC_SERVER_IP}:${process.env.EXPO_PUBLIC_SERVER_PORT}/events/${eventId}`,
+        {
+          headers: {
+            'Authorization': `Bearer ${user.authToken}`,
+            'Content-Type': 'application/json'
+          },
+        }
+      );
+      return response.data.data;
     } catch (err) {
-        console.error('Error fetching events:', err);
-        return null;
+      console.error('Error fetching events:', err);
+      return null;
     }
-};
+  };
 
   const sortEvents = (order: 'mostRecent' | 'leastRecent') => {
-      const sortedEvents = [...events].sort((a, b) => {
-          const comparison = new Date(b.date).getTime() - new Date(a.date).getTime();
-          return order === 'mostRecent' ? comparison : -comparison;
-      });
-      setEvents(sortedEvents);
+    const sortedEvents = [...events].sort((a, b) => {
+      const comparison = new Date(b.date).getTime() - new Date(a.date).getTime();
+      return order === 'mostRecent' ? comparison : -comparison;
+    });
+    setEvents(sortedEvents);
   };
 
   const handleSortChange = (itemValue: string) => {
-      if (itemValue === 'mostRecent' || itemValue === 'leastRecent') {
-          setSortOrder(itemValue);
-      }
+    if (itemValue === 'mostRecent' || itemValue === 'leastRecent') {
+      setSortOrder(itemValue);
+    }
   };
 
   const eventMap = events.map((event) => (
     <View key={event._id} style={styles.eventCard}>
-        <View style={styles.cardHeader}>
-            <Text style={styles.eventName}>{event.name}</Text>
+      <View style={styles.cardHeader}>
+        <Text style={styles.eventName}>{event.name}</Text>
+      </View>
+      <View style={styles.cardBody}>
+        <View style={styles.infoRow}>
+          <Text style={styles.infoLabel}>Date:</Text>
+          <Text style={styles.infoValue}>
+            {new Date(event.date).toLocaleDateString()}
+          </Text>
         </View>
-        <View style={styles.cardBody}>
-            <View style={styles.infoRow}>
-                <Text style={styles.infoLabel}>Date:</Text>
-                <Text style={styles.infoValue}>
-                    {new Date(event.date).toLocaleDateString()}
-                </Text>
-            </View>
-            <View style={styles.infoRow}>
-                <Text style={styles.infoLabel}>Time:</Text>
-                <Text style={styles.infoValue}>
-                    {new Date(event.date).toLocaleTimeString()}
-                </Text>
-            </View>
-            <View style={styles.infoRow}>
-                <Text style={styles.infoLabel}>ID:</Text>
-                <Text style={styles.infoValue}>{event._id}</Text>
-            </View>
+        <View style={styles.infoRow}>
+          <Text style={styles.infoLabel}>Time:</Text>
+          <Text style={styles.infoValue}>
+            {new Date(event.date).toLocaleTimeString()}
+          </Text>
         </View>
+        <View style={styles.infoRow}>
+          <Text style={styles.infoLabel}>ID:</Text>
+          <Text style={styles.infoValue}>{event._id}</Text>
+        </View>
+      </View>
     </View>)
   )
 
   return (
-      <View style={styles.container}>
-        <Text style={styles.header}>Passport</Text>
-        <View style={styles.dropDownLabelContainer}>
-          <Image
-            source={{uri: "https://icons.veryicon.com/png/o/miscellaneous/partme/filter-147.png"}}
-            style={styles.filterImage}
-          />
-          <Text style={styles.filterText}>Filter By: </Text>
-          <DropDownPicker style={styles.dropDown}
-                  textStyle={styles.dropDownText}
-                  open={open}
-                  value={sortOrder}
-                  items={items}
-                  setOpen={setOpen}
-                  setValue={setSortOrder}
-                  setItems={setItems}
-                  containerStyle={{borderRadius: 36, width: 140}}
-            />
-         </View>
-        <ScrollView>
-          <View style={styles.eventList}>
-          {events.length > 0 ? (
-              eventMap
-            ) : (
-              <Text style={styles.noEventsText}>No events found in your passport</Text>
-            )}
-          </View>
-          </ScrollView>
+    <View style={styles.container}>
+      <Text style={styles.header}>Passport</Text>
+      <View style={styles.dropDownLabelContainer}>
+        <Image
+          source={{ uri: "https://icons.veryicon.com/png/o/miscellaneous/partme/filter-147.png" }}
+          style={styles.filterImage}
+        />
+        <Text style={styles.filterText}>Filter By: </Text>
+        <DropDownPicker style={styles.dropDown}
+          textStyle={styles.dropDownText}
+          open={open}
+          value={sortOrder}
+          items={items}
+          setOpen={setOpen}
+          setValue={setSortOrder}
+          setItems={setItems}
+          containerStyle={{ borderRadius: 36, width: 140 }}
+        />
       </View>
+      <ScrollView>
+        <View style={styles.eventList}>
+          {events.length > 0 ? (
+            eventMap
+          ) : (
+            <Text style={styles.noEventsText}>No events found in your passport</Text>
+          )}
+        </View>
+      </ScrollView>
+    </View>
   );
 }
 
 const styles = StyleSheet.create({
   container: {
-      flex: 1,
-      padding: 16,
-      backgroundColor: '#ffffff',
+    flex: 1,
+    padding: "5%",
+    paddingTop: "10%",
+    backgroundColor: '#ffffff',
   },
   pickerContainer: {
     marginVertical: 10,
@@ -248,7 +247,7 @@ const styles = StyleSheet.create({
     flexDirection: 'row',         // Align items in a row (horizontal)
     justifyContent: 'center',    // Center the components horizontally
     alignItems: 'center',        // Center the components vertically
-    marginLeft: 10,  
+    marginLeft: 10,
     marginBottom: 10,         // Optional: add margin around the container
   },
   dropDown: {
